Fixing the ProductFilter to filter recalled products
I am assuming that the requirement is to find the issues with existing functionality.

Given the existing APIs, it seems like the main issue with existing code is that we do not filter out recalled products from the GET Products API. For this the filter class had to be implemented, filter class queries the recalled products from the memory database and filters products using that information. The result of products should not contain any recalled product, unit tests are added for this scenarios as well.